+++ src/ocaml_protoc_plugin/ocaml_protoc_plugin.ml
@@ -2,6 +2,7 @@
 module Serialize = Serialize
 module Deserialize = Deserialize
 module Spec = Spec
+module Runtime = Runtime
 (**/**)
 
 module Reader = Reader
@@ -11,6 +12,7 @@ module Result = Result
 module Extensions = Extensions
 
 (**/**)
+
 let test () =
   Writer.Test.test ();
   Serialize.Test.test ();