+++ src/ocaml_protoc_plugin/runtime.ml
@@ -1,0 +1,10 @@
+module Runtime' = struct
+  module Serialize = Serialize
+  module Deserialize = Deserialize
+  module Spec = Spec
+  module Result = Result
+  module Service = Service
+  module Extensions = Extensions
+  module Reader = Reader
+  module Writer = Writer
+end